move planets modal data js file from public to data subdirectory in public
