+++ public/js/manipulateDomPlanets.js
@@ -1,5 +1,5 @@
 import { replaceModalTitles as replaceModalTitles } from "/replaceModalTitles.js";
-import { modalKeyTitles } from "/modalData.js";
+import { modalKeyTitles } from "/data/modalData.js";
 
 //event listener needs to be added as not all DOM content referenced will be loaded immediately
 document.addEventListener("DOMContentLoaded", (event) => {
